Move state models into core
