+++ apps/web/src/app/core/index.ts
@@ -2,4 +2,4 @@ export * from "./core.module";
 export * from "./event.service";
 export * from "./policy-list.service";
 export * from "./router.service";
-export * from "./state.service";
+export * from "./state/state.service";
