+++ apps/web/src/app/core/state/global-state.ts
